Conflicts:
	WebContent/WEB-INF/views/customer/customerSales.jsp

+++ src/com/pogo/controller/MasterMasterController.java
@@ -25,6 +25,10 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.pogo.bean.AddActionBean;
+import com.pogo.bean.AddPlanBean;
+import com.pogo.bean.CompanyProfileBean;
+import com.pogo.bean.CompetitiorsProfileBean;
 import com.pogo.bean.CountryBean;
 import com.pogo.bean.CurrencyBean;
 import com.pogo.bean.CustomerLevelsBean;
@@ -39,6 +43,7 @@ import com.pogo.bean.ServiceProviderBean;
 
 import com.pogo.bean.StateBean;
 import com.pogo.bean.TeamSegmentBean;
+import com.pogo.model.AddAction;
 import com.pogo.model.Country;
 import com.pogo.model.CustomerLevels;
 import com.pogo.model.District;
@@ -51,6 +56,7 @@ import com.pogo.service.MasterMastersService;
 public class MasterMasterController {
 	@Autowired
 	private MasterMastersService masterMastersService;
+	
 	@RequestMapping(value="/customerLevels",method = RequestMethod.GET)
 	public ModelAndView getcustomerLevels( @ModelAttribute("command") CustomerLevelsBean customerlevel,HttpServletRequest request,BindingResult result ){
 		System.out.println("in customerlevels  method");
@@ -142,28 +148,15 @@ public class MasterMasterController {
 	@RequestMapping(value="/country",method = RequestMethod.GET)
 	public ModelAndView getcountry( @ModelAttribute("command") CountryBean country,HttpServletRequest request,BindingResult result ){
 		System.out.println("inside country  method");
+		
+		List<CountryBean> countrybean=masterMastersService.countryList();
+		
 		Map<String, Object> model = new HashMap<String, Object>();
-		model.put("countryList",  prepareCountryListofBean(masterMastersService.countryList()));
+		model.put("countryList", countrybean);
 		System.out.println("***************************************** inside country list ****************************");
 	return new ModelAndView("getcountry",model);
 	}
 	
-	private List<CountryBean> prepareCountryListofBean(List<Country> prodel){
-		List<CountryBean> beans = null;
-		if(prodel != null && !prodel.isEmpty()){
-			beans = new ArrayList<CountryBean>();
-			CountryBean bean = null;
-			for(Country pro : prodel){
-				bean = new CountryBean();
-				//System.out.println(bean);
-				bean.setCountryId(pro.getCountryId());
-				bean.setCountry(pro.getCountry());
-				
-				beans.add(bean);
-			}
-		}
-		return beans;
-	}
 	
 	@RequestMapping(value="addcountry",method=RequestMethod.POST)
 	@ResponseBody
@@ -177,11 +170,26 @@ public class MasterMasterController {
 	}
 		
 	@RequestMapping(value = "deletecountry", method = RequestMethod.GET)
-	public ModelAndView deleteCountry(@RequestParam("countryId") Integer id) {
-		masterMastersService.deleteCountry(id);
+	
+	public ModelAndView deleteCountry(@RequestParam("countryId") Integer id,Model m) {
+		
+		
+		List<StateBean> bean=masterMastersService.getStateByCountryId(id);
+		
+		System.out.println("<><><><><><><><>   &*&*&*&*&*&*&*    *&*&*&*&    "+bean.isEmpty());
+		//
+		if(bean.isEmpty()==false){
+			m.addAttribute("notdeleting", "Please delete state first");
+			}else{
+				masterMastersService.deleteCountry(id);
+			}
+		List<CountryBean> countrybean=masterMastersService.countryList();
 		Map<String, Object> model= new HashMap<String,Object>();
 		List<CountryBean> list=new ArrayList<CountryBean>();
-		model.put("countryList",  prepareCountryListofBean(masterMastersService.countryList()));
+		model.put("countryList",  countrybean);
+		
+			
+			
 		return new ModelAndView("getcountry",model);
 		
 		
@@ -261,8 +269,16 @@ public class MasterMasterController {
 	}
 	
 	@RequestMapping(value = "deletestate", method = RequestMethod.GET)
-	public ModelAndView deleteState(@RequestParam("stateId") Integer id) {
-		masterMastersService.deleteState(id);
+	public ModelAndView deleteState(@RequestParam("stateId") Integer id , Model m) {
+		List<DistrictBean> bean = masterMastersService.getDistrictByStateIdAndcountryId(id);
+		System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@ "+bean.isEmpty());
+		if(bean.isEmpty()==false){
+			m.addAttribute("notdeleting", "Please delete district first");
+			
+		}else {
+			masterMastersService.deleteState(id);
+		}
+		
 		Map<String, Object> model= new HashMap<String,Object>();
 		List<StateBean> list=new ArrayList<StateBean>();
 		model.put("stateList",  prepareStateListofBean(masterMastersService.stateList()));
@@ -344,8 +360,18 @@ public class MasterMasterController {
 		masterMastersService.addDistrict(poref);
 	}
 	@RequestMapping(value = "deletedistrict", method = RequestMethod.GET)
-	public ModelAndView deleteDistrict(@RequestParam("districtId") Integer id) {
-		masterMastersService.deleteDistrict(id);
+	public ModelAndView deleteDistrict(@RequestParam("districtId") Integer id , Model m) {
+		
+		List<LocationBean> bean = masterMastersService.getLocationDetails();
+		System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@ "+bean.isEmpty());
+		if(bean.isEmpty()==false){
+			m.addAttribute("notdeleting", "Please delete district first");
+			
+		}else {
+			masterMastersService.deleteDistrict(id);
+		}
+		
+		
 		Map<String, Object> model= new HashMap<String,Object>();
 		List<DistrictBean> list=new ArrayList<DistrictBean>();
 		model.put("districtList",  prepareDistrictListofBean(masterMastersService.districtList()));
@@ -837,4 +863,57 @@ public class MasterMasterController {
 	
 	}
 	
+	
+	//satyendra method
+	@RequestMapping(value="/addplanAction",method=RequestMethod.GET)
+	public ModelAndView getPlanAction(@ModelAttribute("command") DistrictBean district,HttpServletRequest request,BindingResult result)
+   {
+		
+		System.out.println("inside add plan action");
+		List<AddAction> list = new ArrayList<AddAction>();
+		list = masterMastersService.actionPlanList();
+		Map<String , Object> model = new HashMap<String,Object>();
+		model.put("actionplanlist", list);
+		//System.out.println("***************************************** inside action plan ****************************");
+	
+		return new ModelAndView("addplanAction",model);
+	}
+	
+	@RequestMapping(value="addactionplan",method=RequestMethod.POST)
+	@ResponseBody
+	public void addActionPlan(@RequestBody String json,Model model) throws IOException{
+	System.out.println("********************inside add Action Plan method **************\n"+json);
+		ObjectMapper mapper=new ObjectMapper();
+		AddActionBean poref=mapper.readValue(json, AddActionBean.class);
+		AddActionBean poref1=new AddActionBean();
+		poref1.setAction(poref.getAction());
+		
+		masterMastersService.addActionPlan(poref1);
+	}
+	@RequestMapping(value="editforactionplan" , method=RequestMethod.POST)
+	public String EditForActionPlan(@RequestParam("id") int id,  Model model) {
+		AddActionBean addobj=masterMastersService.EditForActionPlan(id);
+		model.addAttribute("planname", addobj);
+		
+		return "addplanAction";
+	}
+	
+	//call the page for add actrion
+	@RequestMapping(value = "/addaction", method = RequestMethod.GET)
+	public ModelAndView getAction() {
+		
+     return new ModelAndView("addaction");
+	}
+	//save the add actin page
+	@RequestMapping(value = "saveaddaction", method = RequestMethod.POST)
+	@ResponseBody
+	public void saveAddAction(@RequestBody String json, Model model) throws IOException {
+		System.out.println("i m in controller Add action data   \n" + json);
+		ObjectMapper mapper = new ObjectMapper();
+		AddPlanBean planbean = mapper.readValue(json, AddPlanBean.class);
+
+		AddPlanBean pbean = new AddPlanBean();
+
+		masterMastersService.saveAddAction(planbean);
+	}
 }
